refactor: :art: DRY types
moved common types to new core module
