+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod core;
 pub mod json;
 pub mod yaml;
